% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpitopeFinder.R
\name{Epitope_finder}
\alias{Epitope_finder}
\title{Fahrenheit conversion}
\usage{
Epitope_finder(Hydrophobicity_predictor_output, rm_hydrophobic_regions = FALSE)
}
\arguments{
\item{Hydrophobicity_predictor_output}{The output list generated from the Hydrophobicity_predictor_output function}

\item{rm_hydrophobic_regions}{Logical argument to specify if the epitopes found inside hydrophobic regions should be included in the resulting dataframe}
}
\value{
Dataframe consisting of epitopes, and their location, found in the query protein sequence as well as metadata about the original epitope
}
\description{
Maps epitopes contained in an IEDB export file to the query protein, by taking the output of the Hydrophobicity_predictor function as input, and selecting the desired IEDB export .csv file from the prompt.
}
\examples{
Protein_name_epitope_dataframe <- Epitope_finder(Protein_name_hydrophobicity_output, rm_hydrophobic_regions = FALSE)
}
