% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TM_predictor.R
\name{Hydrophobicity_predictor}
\alias{Hydrophobicity_predictor}
\title{Hydrophobic region predictor}
\usage{
Hydrophobicity_predictor(length = 20)
}
\arguments{
\item{length}{is the desired size of the rolling average}
}
\value{
A list of several elements including the hydrophobicity plot, the hydrophobicity dataframe, the input sequence and a secondary dataframe used in downstream analysis
}
\description{
Predict hydrophobic region of a protein by selecting a protein sequence file from the prompt
}
\examples{
Protein_name_hydrophobicity_output <- Hydrophobicity_predictor(20)
Protein_name_hydrophobicity_plot <- Hydrophobicity_predictor(20)[[1]];
Protein_name_hydrophobicity_df <- Hydrophobicity_predictor(20)[[2]];
Protein_name_input_sequence <- Hydrophobicity_predictor(20)[[3]];
}
